Consumed analysis: add 'consumable' class attribute. Patch by chris.wailes@gmail.com
Adds the 'consumable' attribute that can be attached to classes.  This replaces
the previous method of scanning a class's methods to see if any of them have
consumed analysis attributes attached to them.  If consumed analysis attributes
are attached to methods of a class that isn't marked 'consumable' a warning
is generated.

git-svn-id: https://llvm.org/svn/llvm-project@189702 91177308-0d34-0410-b5e6-96231b3b80d8