[PECOFF] Add a field for machine type to PECOFFLinkingContext.
So that we can determine what the target architecture is. Adding this
field does not mean that we are going to support non-i386 architectures
soon; there are many things to do to support them, and I'm focusing on
i386 now. But this is the first step toward multi architecture support.

git-svn-id: https://llvm.org/svn/llvm-project@190627 91177308-0d34-0410-b5e6-96231b3b80d8