<rdar://problem/14146606>
Fixed an issue where environment variables that contained special characters '$' and '#' would hose up the GDB server packet. We now use the QEnvironmentHexEncoded packet that has existed for a long time when we need to. Also added code that will stop sending the QEnvironmentHexEncoded and QEnvironment packets if they aren't supported.



git-svn-id: https://llvm.org/svn/llvm-project@192373 91177308-0d34-0410-b5e6-96231b3b80d8