Change the clang driver with the use of -no-integrated-as for darwin to use the -Q flag to the as(1) assembler driver.
We will soon be switching the darwin as(1) assembler driver to call clang(1)
and use the intergated assembler by default.  To do this and still support
clang(1)'s -no-integrated-as flag, when clang(1) runs the as(1) assembler
driver and -no-integrated-as is used it needs to pass the -Q flag to as(1)
so it uses its GNU based assembler, and not turn around and call clag(1)'s
integrated assembler. 

rdar://15495921


git-svn-id: https://llvm.org/svn/llvm-project@195054 91177308-0d34-0410-b5e6-96231b3b80d8